backend/requirements.txt
langchain
openai
qdrant_client
transformers
python-dotenv
fastapi
langgraph
uvicorn
sentence_transformers
redis
langchain-community